Feature/3d and 2d (#158)
* feat: add 3d in cach

* feat: 2d and 3d in cache

* chore: delete raylib log

* chore: delete unused parts

* feat: basic merge of 2d and 3d

* fix: player in the correct pos

* feat: clang format

* style: snake_case to camelCase

* fix: compile error in windows with variant

* chore: delete unused includes

* style: snake_case to camelCase

* fix: build error because of M_PI

* chore: delete magic value

* feat: Track large files with Git LFS

* feat: add obj file

---------

+++ libs/engine/includes/Components/Position.hpp
@@ -38,7 +38,7 @@ struct Position : engine::Serializable<Position> {
     /**
      * @brief The z-coordinate of the position.
      */
-    int _z = 0;
+    float _z = 0;
 
     /**
      * @brief Default constructor for the Position structure.
@@ -58,7 +58,7 @@ struct Position : engine::Serializable<Position> {
      * @param y The y-coordinate of the position.
      * @param z The z-coordinate of the position.
      */
-    Position(float x, float y, int z = 0)
+    Position(float x, float y, float z = 0)
         : _x(x)
         , _y(y)
         , _z(z)
